# remote-write: buffer struct instead of interface to reduce garbage-collection