This re-uses the data structure and benchmark improvement from #9904.

Holding a slice of concrete structs dramatically reduces the number of individual objects allocated, allowing sends to run a bit faster.

Previously `BenchmarkSampleDelivery` spent a lot of effort checking each sample had arrived, so was largely showing the performance of test-only code. Rename to `BenchmarkSampleSend`. Also increase the number of shards to be more realistic for a large workload.

Results:
```
name          old time/op    new time/op    delta
SampleSend-4    2.04ms ± 6%    1.73ms ± 2%  -15.23%  (p=0.008 n=5+5)

name          old alloc/op   new alloc/op   delta
SampleSend-4     835kB ± 0%      47kB ± 0%  -94.34%  (p=0.008 n=5+5)

name          old allocs/op  new allocs/op  delta
SampleSend-4     10.1k ± 0%      0.1k ± 0%  -98.62%  (p=0.008 n=5+5)
```